local-agent/requirements.txt
# Local Agent Dependencies
# Python 3.11+ required

# LLM and AI
exllamav2>=0.2.0         # ExLlamaV2 for running Qwen models
torch>=2.0.0             # PyTorch for model inference
transformers>=4.35.0     # Hugging Face transformers

# API Server
fastapi>=0.104.0         # FastAPI framework
uvicorn[standard]>=0.24.0  # ASGI server
pydantic>=2.5.0          # Data validation

# Vector Store (optional but recommended)
chromadb>=0.4.18         # Vector database for semantic search
sentence-transformers>=2.2.2  # For generating embeddings

# Utilities
aiofiles>=23.2.1         # Async file operations
python-dotenv>=1.0.0     # Environment variables
huggingface-hub>=0.19.0  # For downloading models
